refactor: retain private name for now
